+++ packages/apollo-angular/ng-package.js
@@ -7,4 +7,7 @@ module.exports = {
     flatModuleFile: 'ng.apollo',
     umdModuleIds,
   },
+  whitelistedNonPeerDependencies: [
+    "semver"
+  ]
 };