+++ packages/apollo-angular/schematics/install/index.ts
@@ -1,5 +1,5 @@
 import { dirname } from 'path';
-
+import { satisfies } from 'semver';
 import {
   apply,
   chain,
@@ -13,6 +13,8 @@ import {
 } from '@angular-devkit/schematics';
 import { NodePackageInstallTask } from '@angular-devkit/schematics/tasks';
 import { getAppModulePath } from '@schematics/angular/utility/ng-ast-utils';
+import { resolve } from '@angular-devkit/core/node';
+import { tags, terminal } from '@angular-devkit/core';
 
 import { getJsonFile, getMainPath } from '../utils';
 import { Schema } from './schema';
@@ -20,11 +22,61 @@ import { addModuleImportToRootModule } from '../utils/ast';
 
 export default function install(options: Schema): Rule {
   return chain([
+    assertTypescriptVersion(),
     addDependencies(),
     addSetupFiles(options),
     importSetupModule(options),
     importHttpClientModule(options),
   ]);
+}
+
+export const dependenciesMap: Record<string, string> = {
+  'apollo-angular': '^1.3.0',
+  'apollo-angular-link-http': '^1.2.0',
+  'apollo-link': '^1.2.0',
+  'apollo-client': '^2.4.0',
+  'apollo-cache-inmemory': '^1.2.0',
+  'graphql-tag': '^2.9.2',
+  'graphql': '^0.13.2'
+};
+
+function assertTypescriptVersion() {
+  return (host: Tree) => {
+    const allowed = '>=2.8.0';
+    let tsVersion: string;
+
+    try {
+      const resolveOptions = {
+        basedir: host.root.path,
+        checkGlobal: false,
+        checkLocal: true,
+      };
+      tsVersion = require(resolve('typescript', resolveOptions)).version;
+    } catch {
+      console.error(terminal.bold(terminal.red(tags.stripIndents`
+        Versions of typescript could not be determined.
+        The most common reason for this is a broken npm install.
+        Please make sure your package.json contains typescript in
+        devDependencies, then delete node_modules and package-lock.json (if you have one) and
+        run npm install again.
+      `)));
+      process.exit(2);
+
+      return;
+    }
+
+    if (!satisfies(tsVersion, allowed)) {
+      // First line of warning looks weird being split in two, disable tslint for it.
+      console.error((terminal.yellow('\n' + tags.stripIndent`
+        typescript@'${allowed}' is required but ${tsVersion} was found instead.
+        Using this version can result in undefined behaviour and difficult to debug problems.
+        Please run the following command to install a compatible version of TypeScript.
+            npm install typescript@'${allowed}'
+      ` + '\n')));
+    }
+
+    return host;
+  }
 }
 
 /**
@@ -37,22 +89,18 @@ function addDependencies() {
     const packageJsonPath = 'package.json';
     const packageJson = getJsonFile(host, packageJsonPath);
 
+    // TODO: check if TypeScript >= 2.8.0
+
     packageJson.dependencies = packageJson.dependencies || {};
-    const dependenciesToAdd = [
-      'apollo-angular',
-      'apollo-angular-link-http',
-      'apollo-link',
-      'apollo-client',
-      'apollo-cache-inmemory',
-      'graphql-tag',
-      'graphql',
-    ];
-    dependenciesToAdd.forEach(dependency => {
-      if (!packageJson.dependencies[dependency]) {
-        // target the 'latest' tag of every package
-        packageJson.dependencies[dependency] = 'latest';
+
+    for (const dependency in dependenciesMap) {
+      if (dependenciesMap.hasOwnProperty(dependency)) {
+        const version = dependenciesMap[dependency];
+        if (!packageJson.dependencies[dependency]) {
+          packageJson.dependencies[dependency] = version;
+        }
       }
-    });
+    }
 
     // save the changed file
     host.overwrite(packageJsonPath, JSON.stringify(packageJson, null, 2));