+++ packages/apollo-angular/schematics/jest.config.js
@@ -1,0 +1,7 @@
+const jestConfig = require('../../../jest.config');
+
+module.exports = Object.assign(jestConfig(), {
+  transform: {
+    '.ts': 'ts-jest',
+  },
+});