+++ packages/apollo-angular/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### vNEXT
 
-- Make `Subscription` generic
+- Make `Subscription` generic [PR #778](https://github.com/apollographql/apollo-angular/pull/778)
+- Schematics (`ng add apollo-angular`) [PR #779](https://github.com/apollographql/apollo-angular/pull/779), [PR #780](https://github.com/apollographql/apollo-angular/pull/780)
 
 ### v1.2.0